+++ app/controllers/ai_credentials_controller.rb
@@ -2,6 +2,11 @@ class AiCredentialsController < CredentialsController
   self.credential_class = AiCredential
   self.validation_job = AiCredentialValidationJob
 
+  def show
+    super
+    @credential.refresh_models_async
+  end
+
   private
 
   def default_provider