+++ app/jobs/ai_model_catalog_timeout_job.rb
@@ -1,0 +1,7 @@
+class AiModelCatalogTimeoutJob < ApplicationJob
+  queue_as :timeouts
+
+  def perform(run)
+    run.timeout!
+  end
+end